+++ opm/parser/eclipse/EclipseState/Schedule/TimeMap.hpp
@@ -41,12 +41,12 @@ namespace Opm {
         int numTimesteps() const { return m_timeList.size() - 1; }
         double getTotalTime() const;
         /// Return the date and time where a given time step starts.
-        boost::posix_time::ptime getStartTime(int tStepIdx) const
+        boost::posix_time::ptime getStartTime(size_t tStepIdx) const
         { return m_timeList[tStepIdx]; }
         /// Return the period of time in seconds which passed between the start of the simulation and a given time step.
-        double getTimePassedUntil(int tStepIdx) const;
+        double getTimePassedUntil(size_t tStepIdx) const;
         /// Return the length of a given time step in seconds.
-        double getTimeStepLength(int tStepIdx) const;
+        double getTimeStepLength(size_t tStepIdx) const;
         static boost::posix_time::ptime timeFromEclipse(DeckRecordConstPtr dateRecord);
         static boost::posix_time::ptime timeFromEclipse(int day , const std::string& month, int year, const std::string& eclipseTimeString = "00:00:00.000");
         static boost::posix_time::time_duration dayTimeFromEclipse(const std::string& eclipseTimeString);
